@@ -1,32 +1,13 @@
-// by womendi
-import Actions from './actions'
-window.api = require('./config.js')
-// 生成 mutations 方法
-function generate (name) {
-  return function (state, value) {
-    state[name] = value || state[name]
-  }
-}
+import Vue from 'vue';
+import Vuex from 'vuex';
+Vue.use(Vuex);
 
-const options = {
+import commonModule from './commonModule/';
+import donateModule from './donateModule/';
 
-  // 多页面共享数据
-  state: {
-    isLogin: 0
-  },
-
-  // 操作
-  actions: Actions,
-
-  // 数据变更
-  mutations: {
-
-  }
-}
-
-// 为每个 state 字段生成对应的 mutations 方法
-Object.keys(options.state).forEach((key) => {
-  options.mutations[key] = generate(key)
-})
-
-export default options
+module.exports = new Vuex.Store({
+    modules: {
+        commonModule,
+        donateModule
+    }
+});
